# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DOWAR API",
        "description": "API Donasi & Feedback System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Lihat semua program donasi (Publik)",
                "operationId": "f63bfb37edf1acfd021c50961b734ec0",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Buat program donasi baru (Hanya Admin)",
                "operationId": "272d9b28b56c8e18d5ef4c7440a05def",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Bantuan Pakaian Malang"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Donasi pakaian layak pakai."
                                    },
                                    "target_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-25"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign Dibuat"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Detail program donasi spesifik",
                "operationId": "0f74cc63090e3f9d520d77776391f420",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Update program donasi (Hanya Admin)",
                "operationId": "3e659e761ad8967d64e5beab16070c9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Bantuan Pakaian Malang V2"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Update deskripsi."
                                    },
                                    "target_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-01"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Hapus program donasi (Hanya Admin)",
                "operationId": "c640e140d1bee392e2477ab304b63927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/distributions": {
            "get": {
                "tags": [
                    "Distributions"
                ],
                "summary": "Lihat riwayat penyaluran barang",
                "operationId": "1c50a7e7560b6a28a8dc55c38f7ed1c1",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil riwayat distribusi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Distributions"
                ],
                "summary": "Catat pengiriman barang donasi baru",
                "operationId": "cf0677ca3a834ee6995836e6cc7118b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "beneficiary_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Terkirim"
                                    },
                                    "delivery_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-05-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Distribusi Berhasil Dicatat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Lihat riwayat transaksi donasi",
                "operationId": "efd463fc838a5d0b3edcc89ecc8374ac",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data donasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Buat transaksi donasi baru",
                "operationId": "a83d56537f533c01a1daf0abefdec913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "QRIS"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donasi Berhasil Diproses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}